pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slipdeck"
version = "0.1.0"
authors = [
    {name = "Star Three LLC"},
]
description = "Quickly generate professional packing slips for TCGPlayer and eBay orders."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "rich==13.9.4",
    "pydantic==2.10.6",
    "pdfplumber==0.11.5",
    "pypdf2==3.0.1", 
    "python-dotenv==1.0.1",
]

[project.urls]
"Homepage" = "https://github.com/star3llc/slipdeck"
"Bug Tracker" = "https://github.com/star3llc/slipdeck/issues"

[project.scripts]
slipdeck = "slipdeck.cli:app"
